requirements.txt
# Production Dependencies for LucasTeam Telegram Bot
# All versions are pinned for reproducible builds
# Last updated: 2024-01

# ============================================================================
# Core Bot Framework
# ============================================================================
python-telegram-bot==20.7
# Telegram bot API wrapper - provides async interface for Telegram Bot API
# Version 20.7 pinned for API stability and compatibility with our handlers

# ============================================================================
# Database & ORM
# ============================================================================
sqlalchemy==2.0.36
# SQL toolkit and ORM - provides database abstraction layer
# Version 2.0.36 pinned for schema stability and migration compatibility
alembic==1.13.0
# Database migration tool - manages database schema versions
# Version 1.13.0 pinned to ensure compatibility with SQLAlchemy 2.0.36

# ============================================================================
# Configuration & Settings Management
# ============================================================================
pydantic==2.12.0
# Data validation using Python type annotations - validates settings at runtime
pydantic-settings==2.0.0
# Settings management for Pydantic - loads and validates environment variables
python-dotenv==1.0.0
# Environment variable management from .env files - loads config/.env at startup

# ============================================================================
# Logging & Monitoring
# ============================================================================
structlog==23.1.0
# Structured logging for better observability

# ============================================================================
# System Utilities
# ============================================================================
psutil==5.9.6
# System and process utilities - used for process management and monitoring
# Required for safe bot restart functionality

# ============================================================================
# Date & Time Handling
# ============================================================================
pytz==2023.3
# Timezone definitions for accurate time handling across different regions
python-dateutil==2.8.2
# Extensions to the standard datetime module - provides flexible date parsing

# ============================================================================
# Data Formats
# ============================================================================
pyyaml==6.0.3
# YAML parser and emitter - used for reading configuration files
# Version 6.0.3 includes important security fixes and Python 3.14 compatibility

# ============================================================================
# Notes:
# - redis has been removed (not used in codebase - verified in task 3.1.1)
# - All versions are pinned with == for reproducibility
# - For development dependencies, see requirements-dev.txt
# ============================================================================
